pyproject.toml
[tool.poetry]
name = "race"
version = "0.1.0"
description = ""
authors = ["vivek kumar"]
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.0"
motor = "^3.6.0"
gunicorn = "^23.0.0"
PyYAML = "^6.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
python-jose = "^3.5.0"
python-keycloak = "^5.8.1"
requests = "^2.32.5"

[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.24.0"
mypy = "^1.13.0"
requests = "^2.32.0"
httpx = "^0.27.0"

[tool.pytest.ini_options]
testpaths = ["tests"]
norecursedirs = ["data", ".git", ".github"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
